Reference counter trick!
This is sort of an hack. I needed the `self` node not to be copied around
in the system. Since this however happens (and that's totally bad for us!)
I put a simple reference counter inside it.

This fakes the duplication (which actually never happens: the same
instance is returned all over and the refcount gets incremented) and
results in a clean system.

This of course won't be thread-safe until fixed with a mutex.