Bugfix
From topology_test it seems that local node can be dumped on the network
(which is reasonable from a certain point of view), and that we are not
provided with the size when undumping a node.

Also a bugfix in the client. The dictionary could, in principle, check a
NULL client, right?